requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
werkzeug>=3.0.0
